backend/requirements.txt
asgiref==3.5.0
boto3==1.21.42
botocore==1.24.42
cffi==1.15.0
Django==3.2
django-appconf==1.0.5
django-cors-headers==3.11.0
django-filter==21.1
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
jmespath==1.0.0
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9